Back-End-Python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
cryptography>=43.0.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
